+++ lib/src/utils.dart
@@ -3,3 +3,4 @@ import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 
 typedef Dataset<T> = IMap<String, T>;
 typedef QueryFn<T> = Query<T> Function(Query<T> colRef);
+typedef DeleteDocData = Map<String, int>;